Changed the screen UI height flow
Removed the "h-screen" from the DIV holding the (button, Login and progress) which removed the default extra screen scroll.

Added a parent DIV to hold both the Logo and the DIV holding the (button, Login and progress) with class="py-10 space-y-20 h-full". The h-full is to allow the page height to flow with more items to be added. More items added will make the page scrollable. The "space-y-20" is to give a decent space between the logo and the button container/DIV

For the DIV holding the (button, Login and progress), I moved it into another DIV with class="py-20" to make the space distinct.